kernel: do not remove device
Removing the device instance is wrong. The SVEC module has a little
System FPGA where the Application FPGA programmer lives.
The System FPGA is the one that we drive with the SVEC driver
when we program the Application FPGA. So, there is no need to
destroy the SVEC device instance.

Signed-off-by: Federico Vaga <federico.vaga@vaga.pv.it>