hdl: remove DDR SELFREFRESH constraint, not used
The xilinx-generated DDR ip-cores do not use self-refresh (the 'enter self-refresh' signal is tied
to ground). This constraint as such is not needed and many times it creates issues because ISE will
remove the signal and then Translate step will fail because it cannot match the constraint to a
signal.